requirements.txt
Django>=4.1.5,<4.2
psycopg2>=2.9.5,<2.10
python-slugify>=7.0.0,<7.1
Pillow>=9.4,<9.5
uWSGI>=2.0.21,<2.1
